+++ src/cmd/exportAll.go
@@ -1,16 +1,13 @@
 package cmd
 
 import (
-	"bytes"
 	"crypto/tls"
-	"encoding/json"
 	"io/ioutil"
 	"log"
 	"mime"
 	"net/http"
 	"os"
 	"strings"
-	"text/tabwriter"
 
 	"github.com/spf13/cobra"
 )
@@ -121,41 +118,4 @@ func exportApp(appId string, outputDirPath string, format string) {
 		ioutil.WriteFile(exportedFile, body1, 0644)
 		log.Println("Successfully created the export file : " + exportedFile)
 	}
-}
-
-func getAppList() (spIdList []Application) {
-
-	var APPURL = SERVER_CONFIGS.ServerUrl + "/t/" + SERVER_CONFIGS.TenantDomain + "/api/server/v1/applications"
-	var list List
-
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-
-	req, _ := http.NewRequest("GET", APPURL, bytes.NewBuffer(nil))
-	req.Header.Set("Authorization", "Bearer "+SERVER_CONFIGS.Token)
-	req.Header.Set("accept", "*/*")
-	defer req.Body.Close()
-
-	httpClient := &http.Client{}
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		log.Fatalln(err)
-	}
-
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	writer := new(tabwriter.Writer)
-	writer.Init(os.Stdout, 8, 8, 0, '\t', 0)
-	defer writer.Flush()
-
-	err = json.Unmarshal(body, &list)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	resp.Body.Close()
-
-	return list.Applications
 }
